requirements.txt
pillow
numpy
gdown